Add update to modularized script
