+++ modules/bookObject.js
@@ -1,0 +1,5 @@
+const bookData = JSON.parse(localStorage.getItem('storageBooksData'))
+  ? JSON.parse(localStorage.getItem('storageBooksData'))
+  : [];
+
+export default bookData;
